requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
line-bot-sdk==3.5.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
python-multipart==0.0.6
psycopg2-binary==2.9.7
gunicorn==21.2.0